git-tfs-id: [https://tfs.smartcom.software/tfs/DotNET_Project_Collection]$/TPM/ReleaseCandidate/Source;C42778

+++ Module.Persist.TPM/Model/History/HistoricalNoneNego.cs
@@ -7,7 +7,7 @@ namespace Module.Persist.TPM.Model.History
     [AssociatedWith(typeof(NoneNego))]
     public class HistoricalNoneNego : BaseHistoricalEntity<Guid>
     {
-        public int? Discount { get; set; }
+        public double? Discount { get; set; }
         public DateTimeOffset? FromDate { get; set; }
         public DateTimeOffset? ToDate { get; set; }
         public DateTimeOffset? CreateDate { get; set; }
